no more snowflake resolution

+++ TheCurator.Logic/Data/SQLite/CountingChannel.cs
@@ -5,10 +5,10 @@ namespace TheCurator.Logic.Data.SQLite
     public class CountingChannel
     {
         [PrimaryKey]
-        public int ChannelId { get; set; }
+        public long ChannelId { get; set; }
 
         public int Count { get; set; }
 
-        public int LastAuthorId { get; set; }
+        public long LastAuthorId { get; set; }
     }
 }
